+++ src/Middleware/NotFoundMiddleware.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2016-2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2016-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
  */
 
@@ -9,25 +9,51 @@ declare(strict_types=1);
 
 namespace Zend\Expressive\Middleware;
 
+use Fig\Http\Message\StatusCodeInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Zend\Expressive\Handler\NotFoundHandler;
+use Zend\Expressive\Template\TemplateRendererInterface;
 
 class NotFoundMiddleware implements MiddlewareInterface
 {
-    /**
-     * @var NotFoundHandler
-     */
-    private $internalHandler;
+    public const TEMPLATE_DEFAULT = 'error::404';
+    public const LAYOUT_DEFAULT = 'layout::default';
 
     /**
-     * @param NotFoundHandler $internalHandler
+     * @var TemplateRendererInterface
      */
-    public function __construct(NotFoundHandler $internalHandler)
-    {
-        $this->internalHandler = $internalHandler;
+    private $renderer;
+
+    /**
+     * This duplicates the property in StratigilityNotFoundHandler, but is done
+     * to ensure that we have access to the value in the methods we override.
+     *
+     * @var ResponseInterface
+     */
+    protected $responsePrototype;
+
+    /**
+     * @var string
+     */
+    private $template;
+
+    /**
+     * @var string
+     */
+    private $layout;
+
+    public function __construct(
+        ResponseInterface $responsePrototype,
+        TemplateRendererInterface $renderer = null,
+        string $template = self::TEMPLATE_DEFAULT,
+        string $layout = self::LAYOUT_DEFAULT
+    ) {
+        $this->responsePrototype = $responsePrototype;
+        $this->renderer = $renderer;
+        $this->template = $template;
+        $this->layout = $layout;
     }
 
     /**
@@ -38,6 +64,41 @@ class NotFoundMiddleware implements MiddlewareInterface
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler) : ResponseInterface
     {
-        return $this->internalHandler->handle($request);
+        if (! $this->renderer) {
+            return $this->generatePlainTextResponse($request);
+        }
+
+        return $this->generateTemplatedResponse($request);
+    }
+
+    /**
+     * Generates a plain text response indicating the request method and URI.
+     */
+    private function generatePlainTextResponse(ServerRequestInterface $request) : ResponseInterface
+    {
+        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
+        $response->getBody()
+            ->write(sprintf(
+                'Cannot %s %s',
+                $request->getMethod(),
+                (string) $request->getUri()
+            ));
+
+        return $response;
+    }
+
+    /**
+     * Generates a response using a template.
+     *
+     * Template will receive the current request via the "request" variable.
+     */
+    private function generateTemplatedResponse(ServerRequestInterface $request) : ResponseInterface
+    {
+        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
+        $response->getBody()->write(
+            $this->renderer->render($this->template, ['request' => $request, 'layout' => $this->layout])
+        );
+
+        return $response;
     }
 }